move arbitrary implementations to dedicated module
